requirements.txt
huggingface_hub
llama-cpp-python
httpx
pydantic-ai

